normal file colorer use lookup tables instead of palettes
